+++ src/helpers/getHistoryEvents.ts
@@ -1,8 +1,11 @@
 import {
-  govCoreContract,
-  payloadsControllerContract,
-  votingMachineContract,
-} from './contracts';
+  IGovernanceCore_ABI,
+  IPayloadsControllerCore_ABI,
+  IVotingMachineWithProofs_ABI,
+} from '@bgd-labs/aave-address-book';
+import { zeroAddress, zeroHash } from 'viem';
+import { getContractEvents } from 'viem/actions';
+
 import { blockLimit, getEventsBySteps } from './eventsHelpers';
 import { InitEvent, InitEventWithChainId } from './types';
 
@@ -14,14 +17,9 @@ async function getPayloadsCreatedEvents({
   endBlock,
   chainId,
 }: InitEventWithChainId) {
-  const payloadsController = payloadsControllerContract({
-    contractAddress,
-    client,
-  });
-
-  const events = await client.getContractEvents({
-    address: payloadsController.address,
-    abi: payloadsController.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IPayloadsControllerCore_ABI,
     eventName: 'PayloadCreated',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -59,7 +57,16 @@ export async function getPayloadsCreated({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      creator: zeroAddress,
+      payloadId: -1,
+      chainId,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+      payloadsController: contractAddress,
+    },
+  ]);
 }
 
 // proposal created
@@ -69,11 +76,9 @@ async function getProposalCreatedEvents({
   startBlock,
   endBlock,
 }: InitEvent) {
-  const govCore = govCoreContract({ contractAddress, client });
-
-  const events = await client.getContractEvents({
-    address: govCore.address,
-    abi: govCore.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IGovernanceCore_ABI,
     eventName: 'ProposalCreated',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -106,7 +111,13 @@ export async function getProposalCreated({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      proposalId: -1,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+    },
+  ]);
 }
 
 // proposal activate for voting
@@ -116,11 +127,9 @@ async function getProposalActivatedEvents({
   startBlock,
   endBlock,
 }: InitEvent) {
-  const govCore = govCoreContract({ contractAddress, client });
-
-  const events = await client.getContractEvents({
-    address: govCore.address,
-    abi: govCore.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IGovernanceCore_ABI,
     eventName: 'VotingActivated',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -153,7 +162,13 @@ export async function getProposalActivated({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      proposalId: -1,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+    },
+  ]);
 }
 
 // voting activate on VM
@@ -163,11 +178,9 @@ async function getProposalActivatedOnVMEvents({
   startBlock,
   endBlock,
 }: InitEvent) {
-  const votingMachine = votingMachineContract({ contractAddress, client });
-
-  const events = await client.getContractEvents({
-    address: votingMachine.address,
-    abi: votingMachine.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IVotingMachineWithProofs_ABI,
     eventName: 'ProposalVoteStarted',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -200,7 +213,13 @@ export async function getProposalActivatedOnVM({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      proposalId: -1,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+    },
+  ]);
 }
 
 // voting closed on VM and voting results sent
@@ -210,11 +229,9 @@ async function getProposalVotingClosedEvents({
   startBlock,
   endBlock,
 }: InitEvent) {
-  const votingMachine = votingMachineContract({ contractAddress, client });
-
-  const events = await client.getContractEvents({
-    address: votingMachine.address,
-    abi: votingMachine.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IVotingMachineWithProofs_ABI,
     eventName: 'ProposalResultsSent',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -247,7 +264,13 @@ export async function getProposalVotingClosed({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      proposalId: -1,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+    },
+  ]);
 }
 
 // proposal queued
@@ -257,11 +280,9 @@ async function getProposalQueuedEvents({
   startBlock,
   endBlock,
 }: InitEvent) {
-  const govCore = govCoreContract({ contractAddress, client });
-
-  const events = await client.getContractEvents({
-    address: govCore.address,
-    abi: govCore.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IGovernanceCore_ABI,
     eventName: 'ProposalQueued',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -294,7 +315,13 @@ export async function getProposalQueued({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      proposalId: -1,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+    },
+  ]);
 }
 
 // payloads queued
@@ -305,14 +332,9 @@ async function getPayloadsQueuedEvents({
   endBlock,
   chainId,
 }: InitEventWithChainId) {
-  const payloadsController = payloadsControllerContract({
-    contractAddress,
-    client,
-  });
-
-  const events = await client.getContractEvents({
-    address: payloadsController.address,
-    abi: payloadsController.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IPayloadsControllerCore_ABI,
     eventName: 'PayloadQueued',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -349,7 +371,15 @@ export async function getPayloadsQueued({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      payloadId: -1,
+      chainId,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+      payloadsController: contractAddress,
+    },
+  ]);
 }
 
 // payloads executed
@@ -360,14 +390,9 @@ async function getPayloadsExecutedEvents({
   endBlock,
   chainId,
 }: InitEventWithChainId) {
-  const payloadsController = payloadsControllerContract({
-    contractAddress,
-    client,
-  });
-
-  const events = await client.getContractEvents({
-    address: payloadsController.address,
-    abi: payloadsController.abi,
+  const events = await getContractEvents(client, {
+    address: contractAddress,
+    abi: IPayloadsControllerCore_ABI,
     eventName: 'PayloadExecuted',
     fromBlock: BigInt(startBlock),
     toBlock: BigInt(endBlock),
@@ -404,5 +429,13 @@ export async function getPayloadsExecuted({
     });
   };
 
-  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc);
+  return getEventsBySteps(startBlock, endBlock, blockLimit, callbackFunc, [
+    {
+      payloadId: -1,
+      chainId,
+      transactionHash: zeroHash,
+      blockNumber: startBlock,
+      payloadsController: contractAddress,
+    },
+  ]);
 }